+++ git-application/src/main/kotlin/com/xquare/v1servicegit/github/dto/response/GetGithubUserInfoResponse.kt
@@ -1,4 +1,4 @@
-package com.xquare.v1servicegit.github.dto
+package com.xquare.v1servicegit.github.dto.response
 
 data class GetGithubUserInfoResponse(
     val login: String?,
