+++ git-application/src/main/kotlin/com/xquare/v1servicegit/github/port/QueryGithubPort.kt
@@ -1,7 +1,7 @@
 package com.xquare.v1servicegit.github.port
 
-import com.xquare.v1servicegit.github.dto.GetGithubUserInfoResponse
-import com.xquare.v1servicegit.github.dto.TokenResponse
+import com.xquare.v1servicegit.github.dto.response.GetGithubUserInfoResponse
+import com.xquare.v1servicegit.github.dto.response.TokenResponse
 
 interface QueryGithubPort {
     suspend fun getAccessToken(code: String): TokenResponse
