--- git-application/src/main/kotlin/com/xquare/v1servicegit/user/dto/FindUserListResponse.kt
@@ -1,16 +1,0 @@
-package com.xquare.v1servicegit.user.dto
-
-import com.fasterxml.jackson.annotation.JsonProperty
-import java.util.UUID
-
-data class FindUserListResponse(
-    val users: List<FindUserInfoElement>,
-)
-
-data class FindUserInfoElement(
-    val id: UUID,
-    val name: String,
-
-    @JsonProperty("profile_file_name")
-    val profileFileName: String?,
-)
