+++ git-application/src/main/kotlin/com/xquare/v1servicegit/user/port/QueryUserPort.kt
@@ -1,8 +1,8 @@
 package com.xquare.v1servicegit.user.port
 
 import com.xquare.v1servicegit.git.dto.request.FindUserInfoRequest
-import com.xquare.v1servicegit.user.dto.FindUserInfoElement
-import com.xquare.v1servicegit.user.dto.FindUserListResponse
+import com.xquare.v1servicegit.user.dto.response.FindUserListResponse
+import com.xquare.v1servicegit.user.dto.response.FindUserListResponse.FindUserInfoElement
 import java.util.UUID
 
 interface QueryUserPort {
